+++ traffic_ops/ort/atstccfg/cfgfile/hostingdotconfig.go
@@ -21,110 +21,44 @@ package cfgfile
 
 import (
 	"errors"
-	"strconv"
 	"strings"
 
 	"github.com/apache/trafficcontrol/lib/go-atscfg"
-	"github.com/apache/trafficcontrol/lib/go-log"
 	"github.com/apache/trafficcontrol/lib/go-tc"
-	"github.com/apache/trafficcontrol/traffic_ops/ort/atstccfg/config"
-	"github.com/apache/trafficcontrol/traffic_ops/ort/atstccfg/toreq"
 )
 
 const ServerHostingDotConfigMidIncludeInactive = false
 const ServerHostingDotConfigEdgeIncludeInactive = true
 
-func GetConfigFileServerHostingDotConfig(cfg config.TCCfg, serverNameOrID string) (string, error) {
-	// TODO TOAPI add /servers?cdn=1 query param
-	servers, err := toreq.GetServers(cfg)
-	if err != nil {
-		return "", errors.New("getting servers: " + err.Error())
-	}
-
-	server := tc.Server{ID: atscfg.InvalidID}
-	if serverID, err := strconv.Atoi(serverNameOrID); err == nil {
-		for _, toServer := range servers {
-			if toServer.ID == serverID {
-				server = toServer
-				break
-			}
-		}
-	} else {
-		serverName := serverNameOrID
-		for _, toServer := range servers {
-			if toServer.HostName == serverName {
-				server = toServer
-				break
-			}
-		}
-	}
-	if server.ID == atscfg.InvalidID {
-		return "", errors.New("server '" + serverNameOrID + " not found in servers")
-	}
-
-	serverName := tc.CacheName(server.HostName)
-
-	toToolName, toURL, err := toreq.GetTOToolNameAndURLFromTO(cfg)
-	if err != nil {
-		return "", errors.New("getting global parameters: " + err.Error())
-	}
-
-	profileParams, err := toreq.GetProfileParameters(cfg, server.Profile)
-	if err != nil {
-		return "", errors.New("getting profile '" + server.Profile + "' parameters: " + err.Error())
-	}
-	if len(profileParams) == 0 {
-		// The TO endpoint behind toclient.GetParametersByProfileName returns an empty object with a 200, if the Profile doesn't exist.
-		// So we act as though we got a 404 if there are no params, to make ORT behave correctly.
-		return "", config.ErrNotFound
-	}
-
-	fileParams := map[string]string{}
-	for _, param := range profileParams {
-		if param.ConfigFile != atscfg.HostingConfigParamConfigFile {
-			continue
-		}
-		if val, ok := fileParams[param.Name]; ok {
-			log.Errorln("hosting config parameter name '" + param.Name + "' got multiple values - using '" + val + "'")
-			continue
-		}
-		fileParams[param.Name] = param.Value
-	}
-
-	dses, err := toreq.GetCDNDeliveryServices(cfg, server.CDNID)
-	if err != nil {
-		return "", errors.New("getting delivery services: " + err.Error())
-	}
+func GetConfigFileServerHostingDotConfig(toData *TOData) (string, string, error) {
+	fileParams := ParamsToMap(FilterParams(toData.ServerParams, atscfg.HostingConfigParamConfigFile, "", "", ""))
 
 	cdnServers := map[tc.CacheName]tc.Server{}
-	for _, sv := range servers {
-		if sv.CDNID != server.CDNID {
+	for _, sv := range toData.Servers {
+		if sv.CDNID != toData.Server.CDNID {
 			continue
 		}
 		cdnServers[tc.CacheName(sv.HostName)] = sv
 	}
 
-	serverIDs := []int{}
+	serverIDs := map[int]struct{}{}
 	for _, sv := range cdnServers {
-		serverIDs = append(serverIDs, sv.ID)
+		serverIDs[sv.ID] = struct{}{}
 	}
 
-	dsIDs := []int{}
-	for _, ds := range dses {
+	dsIDs := map[int]struct{}{}
+	for _, ds := range toData.DeliveryServices {
 		if ds.ID != nil {
-			dsIDs = append(dsIDs, *ds.ID)
+			dsIDs[*ds.ID] = struct{}{}
 		}
 	}
 
-	dsServers, err := toreq.GetDeliveryServiceServers(cfg, dsIDs, serverIDs)
-	if err != nil {
-		return "", errors.New("getting delivery service servers: " + err.Error())
-	}
+	dsServers := FilterDSS(toData.DeliveryServiceServers, dsIDs, serverIDs)
 
 	dsServerMap := map[int]map[int]struct{}{} // set[dsID][serverID]
 	for _, dss := range dsServers {
 		if dss.Server == nil || dss.DeliveryService == nil {
-			return "", errors.New("deliveryserviceservers returned dss with nil values")
+			return "", "", errors.New("deliveryserviceservers returned dss with nil values")
 		}
 		if _, ok := dsServerMap[*dss.DeliveryService]; !ok {
 			dsServerMap[*dss.DeliveryService] = map[int]struct{}{}
@@ -134,9 +68,9 @@ func GetConfigFileServerHostingDotConfig(cfg config.TCCfg, serverNameOrID string
 
 	hostingDSes := map[tc.DeliveryServiceName]tc.DeliveryServiceNullable{}
 
-	isMid := strings.HasPrefix(server.Type, tc.MidTypePrefix)
+	isMid := strings.HasPrefix(toData.Server.Type, tc.MidTypePrefix)
 
-	for _, ds := range dses {
+	for _, ds := range toData.DeliveryServices {
 		if ds.Active == nil || ds.Type == nil || ds.XMLID == nil || ds.CDNID == nil || ds.ID == nil || ds.OrgServerFQDN == nil {
 			// some DSes have nil origins. I think MSO? TODO: verify
 			continue
@@ -146,7 +80,7 @@ func GetConfigFileServerHostingDotConfig(cfg config.TCCfg, serverNameOrID string
 			continue
 		}
 
-		if *ds.CDNID != server.CDNID {
+		if *ds.CDNID != toData.Server.CDNID {
 			continue
 		}
 
@@ -169,7 +103,7 @@ func GetConfigFileServerHostingDotConfig(cfg config.TCCfg, serverNameOrID string
 				continue
 			}
 
-			if _, ok := dsServerMap[*ds.ID][server.ID]; !ok {
+			if _, ok := dsServerMap[*ds.ID][toData.Server.ID]; !ok {
 				continue
 			}
 		}
@@ -186,6 +120,5 @@ func GetConfigFileServerHostingDotConfig(cfg config.TCCfg, serverNameOrID string
 		origins = append(origins, origin)
 	}
 
-	txt := atscfg.MakeHostingDotConfig(serverName, toToolName, toURL, fileParams, origins)
-	return txt, nil
+	return atscfg.MakeHostingDotConfig(tc.CacheName(toData.Server.HostName), toData.TOToolName, toData.TOURL, fileParams, origins), atscfg.ContentTypeHostingDotConfig, nil
 }